requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
plotly
scikit-learn
statsmodels
openai
langchain
langchain-openai
langchain-experimental
python-dotenv
fpdf2
openpyxl
google-api-python-client
tavily-python
requests
kaggle